move most tabs into separate widget files
